added game_parser and map_parser adn refactored accordingly
